getgems-community-tact/nft-template-in-tact#1
