+++ .gitignore
@@ -1,1 +1,4 @@
 node_modules
+temp
+build
+package-lock.json
