--- .prettierrc.json
@@ -1,5 +1,0 @@
-{
-    "printWidth": 120,
-    "tabWidth": 4,
-    "useTabs": false
-}
