+++ contracts/imports/globals.tact
@@ -1,0 +1,2 @@
+const minTonsForStorage: Int = ton("0.03");
+const gasConsumption: Int = ton("0.03");
