+++ contracts/imports/messages.tact
@@ -1,0 +1,40 @@
+// [Collection Contract] -> Transfer -> OwnershipAssigned -> NFTExcesses
+message(0x5fcc3d14) Transfer { 
+    query_id: Int as uint64;            
+    new_owner: Address; 
+    response_destination: Address; 
+    custom_payload: Cell?; 
+    forward_amount: Int as coins; 
+    forward_payload: Slice as remaining; 
+}
+
+message(0x05138d91) OwnershipAssigned{
+    query_id: Int as uint64;
+    prev_owner: Address;
+    forward_payload: Slice as remaining; 
+}
+
+message(0xd53276db) Excesses {
+    query_id: Int as uint64;
+}
+
+message(0x2fcb26a2) GetStaticData { 
+    query_id: Int as uint64;
+}
+
+message(0x8b771735) ReportStaticData{
+    query_id: Int as uint64;
+    index_id: Int as uint32;
+    collection: Address;
+}
+
+message(0x693d3950) GetRoyaltyParams {
+    query_id: Int as uint64;
+}
+
+message(0xa8cb00ad) ReportRoyaltyParams {
+    query_id: Int as uint64;
+    numerator:  Int as uint16;
+    denominator: Int as uint16;
+    destination: Address;
+}
