+++ package.json
@@ -1,34 +1,25 @@
 {
-  "private": true,
-  "scripts": {
-    "build": "tact --config ./tact.config.json",
-    "test": "jest",
-    "deploy": "ts-node ./sources/contract.deploy.ts",
-    "read": "ts-node ./sources/contract.read.ts"
-  },
-  "dependencies": {
-    "@orbs-network/ton-access": "^2.2.2",
-    "@tact-lang/compiler": "^1.1.3",
-    "@tact-lang/coverage": "^0.0.7",
-    "@tact-lang/deployer": "^0.2.0",
-    "@tact-lang/emulator": "^3.4.1",
-    "@ton-community/sandbox": "^0.11.0",
-    "@ton-community/test-utils": "^0.3.0",
-    "@types/jest": "^29.2.4",
-    "@types/node": "^18.11.14",
-    "@types/qs": "^6.9.7",
-    "base64url": "^3.0.1",
-    "enquirer": "^2.3.6",
-    "jest": "^29.3.1",
-    "open": "^8.4.0",
-    "prando": "^6.0.1",
-    "prettier": "^2.5.1",
-    "qs": "^6.11.0",
-    "ton": "^13.1.0",
-    "ton-core": ">=0.49.0",
-    "ton-crypto": "^3.2.0",
-    "ts-jest": "^29.0.3",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.9.4"
-  }
+    "name": "nft-template",
+    "version": "0.0.1",
+    "private":"true",
+    "scripts": {
+        "start": "blueprint run",
+        "build": "blueprint build",
+        "test": "jest"
+    },
+    "devDependencies": {
+        "@ton-community/blueprint": "^0.12.0",
+        "@ton-community/sandbox": "^0.11.0",
+        "@ton-community/test-utils": "^0.3.0",
+        "@types/jest": "^29.5.0",
+        "@types/node": "^20.2.5",
+        "jest": "^29.5.0",
+        "prettier": "^2.8.6",
+        "ton": "^13.6.0",
+        "ton-core": "^0.52.0",
+        "ton-crypto": "^3.2.0",
+        "ts-jest": "^29.0.5",
+        "ts-node": "^10.9.1",
+        "typescript": "^4.9.5"
+    }
 }
