--- sources/output/sample_NftCollection.constants.fc
@@ -1,2 +1,0 @@
-;; String "meta.json"
-slice __gen_slice_string_7ca4179185a84c879806ec30a1f5fa692451a7fb8dfcf32a967a2cb475822481() asm "B{b5ee9c7241010101000b0000126d6574612e6a736f6ebde57607} B>boc <s PUSHSLICE";
