--- sources/output/sample_NftItem.constants.fc
@@ -1,2 +1,0 @@
-;; String ".json"
-slice __gen_slice_string_2f62f0bd20aef2d2b17f109c6aba1bc3dd7a55db89e349f2659bf698f80ce9c6() asm "B{b5ee9c7241010101000700000a2e6a736f6e9bc05ee0} B>boc <s PUSHSLICE";
