--- sources/output/sample_NftItem.stdlib.fc
@@ -1,196 +1,0 @@
-global (int, slice, int, slice) __tact_context;
-global slice __tact_context_sender;
-global cell __tact_context_sys;
-global int __tact_randomized;
-
-slice __tact_verify_address(slice address) inline {
-    throw_unless(136, address.slice_bits() == 267);
-    var h = address.preload_uint(11);
-    throw_if(137, h == 1279);
-    throw_unless(136, h == 1024);
-    return address;
-}
-
-(slice, slice) __tact_load_address(slice cs) inline {
-    slice raw = cs~load_msg_addr();
-    return (cs, __tact_verify_address(raw));
-}
-
-builder __tact_store_address(builder b, slice address) inline {
-    return b.store_slice(__tact_verify_address(address));
-}
-
-int __tact_my_balance() inline {
-    return pair_first(get_balance());
-}
-
-forall X -> X __tact_not_null(X x) inline {
-    throw_if(128, null?(x)); return x;
-}
-
-(int, slice, int, slice) __tact_context_get() inline {
-    return __tact_context;
-}
-
-builder __tact_store_bool(builder b, int v) inline {
-    return b.store_int(v, 1);
-}
-
-int __tact_slice_eq_bits(slice a, slice b) inline {
-    return equal_slice_bits(a, b);
-}
-
-tuple __tact_string_builder_start(builder b) inline {
-    return tpush(tpush(empty_tuple(), b), null());
-}
-
-tuple __tact_string_builder_start_string() inline {
-    return __tact_string_builder_start(begin_cell());
-}
-
-cell __tact_string_builder_end(tuple builders) inline {
-    (builder b, tuple tail) = uncons(builders);
-    cell c = b.end_cell();
-    while(~ null?(tail)) {
-        (b, tail) = uncons(tail);
-        c = b.store_ref(c).end_cell();
-    }
-    return c;
-}
-
-((tuple), ()) __tact_string_builder_append(tuple builders, slice sc) inline_ref {
-    int sliceRefs = slice_refs(sc);
-    int sliceBits = slice_bits(sc);
-    
-    while((sliceBits > 0) | (sliceRefs > 0)) {
-    
-        ;; Load the current builder
-        (builder b, tuple tail) = uncons(builders);
-        int remBytes = 127 - (builder_bits(b) / 8);
-        int exBytes = sliceBits / 8;
-    
-        ;; Append bits
-        int amount = min(remBytes, exBytes);
-        if (amount > 0) {
-            slice read = sc~load_bits(amount * 8);
-            b = b.store_slice(read);
-        }
-    
-        ;; Update builders
-        builders = cons(b, tail);
-    
-        ;; Check if we need to add a new cell and continue
-        if (exBytes - amount > 0) {
-            var bb = begin_cell();
-            builders = cons(bb, builders);
-            sliceBits = (exBytes - amount) * 8;
-        } elseif (sliceRefs > 0) {
-            sc = sc~load_ref().begin_parse();
-            sliceRefs = slice_refs(sc);
-            sliceBits = slice_bits(sc);
-        } else {
-            sliceBits = 0;
-            sliceRefs = 0;
-        }
-    }
-    
-    return ((builders), ());
-}
-
-slice __tact_int_to_string(int src) inline_ref {
-    var b = begin_cell();
-    if (src < 0) {
-        b = b.store_uint(45, 8);
-        src = - src;
-    }
-    
-    if (src < 1000000000000000000000000000000) {
-        int len = 0;
-        int value = 0;
-        int mult = 1;
-        do {
-            (src, int res) = src.divmod(10);
-            value = value + (res + 48) * mult;
-            mult = mult * 256;
-            len = len + 1;
-        } until (src == 0);
-    
-        b = b.store_uint(value, len * 8);
-    } else {
-        tuple t = empty_tuple();
-        int len = 0;
-        do {
-            int digit = src % 10;
-            t~tpush(digit);
-            len = len + 1;
-            src = src / 10;
-        } until (src == 0);
-    
-        int c = len - 1;
-        repeat(len) {
-            int v = t.at(c);
-            b = b.store_uint(v + 48, 8);
-            c = c - 1;
-        }
-    }
-    return b.end_cell().begin_parse();
-}
-
-() $global_send((int, slice, int, int, cell, cell, cell) $params) impure inline_ref {
-    var (($params'bounce, $params'to, $params'value, $params'mode, $params'body, $params'code, $params'data)) = $params;
-    builder $b = begin_cell();
-    $b = store_int($b, 1, 2);
-    $b = __tact_store_bool($b, $params'bounce);
-    $b = store_int($b, 0, 3);
-    $b = __tact_store_address($b, $params'to);
-    $b = store_coins($b, $params'value);
-    $b = store_int($b, 0, ((((1 + 4) + 4) + 64) + 32));
-    if (( ((~ null?($params'code))) ? (true) : ((~ null?($params'data))) )) {
-        $b = __tact_store_bool($b, true);
-        builder $bc = begin_cell();
-        $bc = __tact_store_bool($bc, false);
-        $bc = __tact_store_bool($bc, false);
-        if ((~ null?($params'code))) {
-            $bc = __tact_store_bool($bc, true);
-            $bc = store_ref($bc, __tact_not_null($params'code));
-        } else {
-            $bc = __tact_store_bool($bc, false);
-        }
-        if ((~ null?($params'data))) {
-            $bc = __tact_store_bool($bc, true);
-            $bc = store_ref($bc, __tact_not_null($params'data));
-        } else {
-            $bc = __tact_store_bool($bc, false);
-        }
-        $bc = __tact_store_bool($bc, false);
-        $b = __tact_store_bool($b, true);
-        $b = store_ref($b, end_cell($bc));
-    } else {
-        $b = __tact_store_bool($b, false);
-    }
-    cell $body = $params'body;
-    if ((~ null?($body))) {
-        $b = __tact_store_bool($b, true);
-        $b = store_ref($b, __tact_not_null($body));
-    } else {
-        $b = __tact_store_bool($b, false);
-    }
-    cell $c = end_cell($b);
-    send_raw_message($c, $params'mode);
-}
-
-slice $Cell$_fun_asSlice(cell $self) impure inline {
-    var ($self) = $self;
-    return begin_parse($self);
-}
-
-int $Context$_fun_readForwardFee((int, slice, int, slice) $self) impure inline_ref {
-    var (($self'bounced, $self'sender, $self'value, $self'raw)) = $self;
-    var (($self'bounced, $self'sender, $self'value, $self'raw)) = $self;
-    slice $sc = $self'raw;
-    $sc~__tact_load_address();
-    $sc~load_coins();
-    $sc~skip_bits(1);
-    $sc~load_coins();
-    return (($sc~load_coins() * 3) / 2);
-}
