--- tact.config.json
@@ -1,10 +1,0 @@
-{
-    "projects": [{
-        "name": "sample",
-        "path": "./sources/contract.tact",
-        "output": "./sources/output",
-        "options": {
-            "external": true  
-        }
-    }]
-}
