+++ wrappers/NftCollection.ts
@@ -1,0 +1,1 @@
+export * from '../build/NftCollection/tact_NftCollection';
