+++ wrappers/NftItem.compile.ts
@@ -1,0 +1,6 @@
+import { CompilerConfig } from '@ton-community/blueprint';
+
+export const compile: CompilerConfig = {
+    lang: 'tact',
+    target: 'contracts/nft_item.tact',
+};
