+++ wrappers/NftItem.ts
@@ -1,0 +1,1 @@
+export * from '../build/NftItem/tact_NftItem';
